Add authentication services

+++ server/PhotonPiano.DataAccess/Interfaces/IGenericRepository.cs
@@ -9,13 +9,13 @@ namespace PhotonPiano.DataAccess.Interfaces
 {
     public interface IGenericRepository<T> where T : class
     {
-        Task<T?> GetByIdAsync(int id);
+        Task<T?> GetByIdAsync(long id);
         Task<IEnumerable<T>> GetAllAsync();
         Task<T?> FindOneAsync(Expression<Func<T, bool>> expression);
         Task<IEnumerable<T>> FindAsync(Expression<Func<T, bool>> expression);
         Task<T> AddAsync(T entity);
         Task UpdateAsync(T entity);
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
         Task<IEnumerable<T>> AddRangeAsync(IEnumerable<T> entities);
         Task UpdateRangeAsync(IEnumerable<T> entities);
         Task DeleteRangeAsync(IEnumerable<T> entities);
